refactor(guid): 将 GuidPage 组件内联到 index.tsx，移除独立文件
